エネミーの仕様変更 敵の効果追加 ステージのタグ追加

+++ Assets/Application/Scripts/Component/EnemyEffect/EnemyEffectBase.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public abstract class EnemyEffectBase : MonoBehaviour {
 	public virtual void MoveEffect (){}
-	public virtual void DamageEffect (int _damage){}
+	public virtual void DamageEffect (float _damage){}
 	public virtual void DeadEffect (){}
 	public virtual void OnTrriger2DEffect (Collider2D _other, int _enemyId){}
 }
