+++ Assets/Application/Scripts/Component/EnemyEffect/EnemyShield.cs
@@ -6,7 +6,7 @@ public class EnemyShield : EnemyEffectBase {
 
 	[SerializeField] private int ProtectTimes = 1;
 
-	public override void DamageEffect (int _damage){
+	public override void DamageEffect (float _damage){
 		if (ProtectTimes > 0) {
 			ProtectTimes--;
 			gameObject.GetComponent<Enemy> ().RecoveryHP (_damage);
